requirements.txt
pandas==1.5.2
prefect==2.7.7
prefect-sqlalchemy==0.2.2
prefect_gcp[cloud_storage]==0.2.3
protobuf==4.21.11
pyarrow==10.0.1
pandas-gbq==0.18.1
psycopg2-binary
#psycopg2-binary==2.9.5
sqlalchemy==1.4.46


prefect
pandas
prefect_gcp
pyarrow
google-cloud-bigquery
prefect_gcp[cloud_storage]==0.2.3